LPS-56854 Moving Comparator to package-private
The current implementation should not be reused outside this component
